chores: be_routes_seperation
