+++ backend/routes/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""라우트 모듈 초기화"""
+
